move commands and everything else to plugin package
