+++ server/plugin/http_autocomplete.go
@@ -1,4 +1,4 @@
-package main
+package plugin
 
 import (
 	"net/http"
