+++ server/plugin/manifest.go
@@ -1,6 +1,6 @@
 // This file is automatically generated. Do not modify it manually.
 
-package main
+package plugin
 
 import (
 	"strings"
